requirements.txt
pytest==7.1.*
pyyaml==6.0.*
munch==2.5.*
click==8.0.4
requests==2.28.*
tqdm==4.64.*
dvc[gs]==2.15.*
